+++ src/util.h
@@ -9,9 +9,9 @@
 
 /* functions provided by util.c */
 int cleanup(void);
-void drawStringAt(int *pageNum, const char *string, int x, int y);
-void drawStringCentered(int *page, const char *str, int startx, int y, int endx);
-int stringWidth(int *page, const char *str);
+void drawStringAt(int16 *pageNum, const char *string, int x, int y);
+void drawStringCentered(int16 *page, const char *str, int startx, int y, int endx);
+int stringWidth(int16 *page, const char *str);
 int my_ltoa(int32 value, char *buf);
 int my_itoa(int value, char *buf);
